Take the Google Maps API key to the client side
This is a bit of a convoluted solution for the configuration and
testing problem. What I was previously trying to do was impossible.

I was trying:
  - to fetch the key asynchronously from the server using the $http
    service;
  - to configure the ngGmap API provider with that key on success.

This is impossible because:
  - services like $http are only available during the run phase;
  - providers are only available during the configuration phase;
  - the configuration phase must finish before the run phase can start.

So this was an instance of the chicken-and-egg problem. The ideal
solution would have been to use Grunt to insert the key at build time,
from a configuration file kept out of VCS. However, the time budget
does not allow for adding another framework.