Convert location component to directive with single attribute
Components do not work as advertised. Converting to a directive was the
most direct way to get this to work. This has been a hard lesson to
avoid angular components.

For improved convenience, I bundled all the required data in a single
"properties" attribute.

I also found that angular-google-maps directives really must have a
zoom level even if the bounds are set, so I replaced the bounds with
the zoom level altogether.