Ditch the nested layout for now, use flat api for attachment writing
A nested layout, with attachment URLs like

/api/stories/<story-id>/url-attachments/<attachment-id>/

would be nicer, but it is complicated to achieve and therefore too
time-consuming for now. So instead, we have this:

/api/url-attachments/<story-id>/

While this is mostly straightforward to implement, there is one catch.
Users should only be able to upload attachments for stories that they
own. The standard permission system of Django REST Framework does not
cater for this scenario, so this will require a little hack.