Make It Work
This required several adjustments:

- PersonSerializer.parent_occupations is now read_only.
Managing parent occupations will require a separate API,
like with the story attachments.

- FormData must be passed a raw DOM element, not an angular
FormController.  addition, FormData.get is not available in all
browsers, so the raw form has to be unaugmented before it is passed
to FormData.

- DRF PUT views require all fields to be present, so I had to insert
temporary "ghost" input fields for the fields that require further work.